requirements.txt
aiopath 
python-dotenv
pyrogram
tgcrypto
uvloop